+++ src/Point.java
@@ -1,13 +1,13 @@
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
     public Point() {
         this.x = 0;
         this.y = 0;
     }
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
